refactor: unite modal image and title into header
